linting + fixing deploy iam roles
Former-commit-id: 8d61bedc9a3f59cb5e84b64d2fabb6263472bfcc

+++ deployment/utilities/iam-deployment.js
@@ -238,7 +238,12 @@ class IAMDeployment extends AWSDeploymentUtilities {
           let role_definitions = this.acquireRoleFile(role_file_name);
 
           let deploy_role_promises = arrayutilities.map(role_definitions, (role_definition) => {
-            return this.destroyRole(role_definition);
+
+						//Technical Debt: this can be way cleaner, can remove role from instance profile then delete
+						if(!role_definition.RoleName === 'DataPipelineDefaultResourceRole'){
+							return this.destroyRole(role_definition);
+						}
+
           });
 
           return Promise.all(deploy_role_promises);
